It is not a swimming pool (#2713)
The translation is swimming pool. I think if plural `"Pools": "Пулы"` then should be `"Pool": "Пул"`. This only make sense. I wanted to change this one the most because it is absolutely incorrect translation that does not make any sense. In Russian Бассейн из exactly Swimming pool. And when you read a translation you think, What!!!? What swimming pool?!!